refactor: create @pnpm/resolve-dependencies
