+++ packages/resolve-dependencies/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@pnpm/resolve-dependencies",
+  "version": "0.0.0",
+  "author": {
+    "name": "Zoltan Kochan",
+    "email": "z@kochan.io",
+    "url": "https://www.kochan.io/"
+  },
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "keywords": [],
+  "license": "MIT",
+  "engines": {
+    "node": ">=6"
+  },
+  "peerDependencies": {
+    "@pnpm/logger": "^2.1.0"
+  },
+  "repository": "https://github.com/pnpm/pnpm/blob/master/packages/resolve-dependencies",
+  "homepage": "https://github.com/pnpm/pnpm/blob/master/packages/resolve-dependencies#readme",
+  "scripts": {
+    "start": "npm run tsc -- --watch",
+    "test": "npm run tsc && npm run lint",
+    "lint": "tslint -c tslint.json --project .",
+    "tsc": "rimraf lib && tsc",
+    "prepublishOnly": "npm run tsc"
+  },
+  "dependencies": {
+    "@pnpm/core-loggers": "0.0.0",
+    "@pnpm/package-requester": "4.1.4",
+    "@pnpm/resolver-base": "1.2.0",
+    "@pnpm/types": "1.8.0",
+    "@pnpm/utils": "0.8.0",
+    "@types/node": "10.11.3",
+    "@types/ramda": "0.25.38",
+    "@types/semver": "5.5.0",
+    "dependency-path": "2.0.0",
+    "package-store": "0.23.4",
+    "path-exists": "3.0.0",
+    "pnpm-install-checks": "1.1.0",
+    "pnpm-shrinkwrap": "7.0.2",
+    "ramda": "0.25.0",
+    "replace-string": "1.1.0",
+    "semver": "5.5.1",
+    "version-selector-type": "2.0.0"
+  },
+  "devDependencies": {
+    "@pnpm/logger": "2.1.0",
+    "rimraf": "2.6.2",
+    "tslint": "5.11.0",
+    "typescript": "3.1.1"
+  }
+}
