+++ packages/resolve-dependencies/src/encodePkgId.ts
