+++ packages/resolve-dependencies/tslint.json
@@ -1,0 +1,44 @@
+{
+  "extends": "tslint:recommended",
+  "rules": {
+    "curly": false,
+    "eofline": false,
+    "align": [true, "parameters"],
+    "class-name": true,
+    "indent": [true, "spaces"],
+    "max-line-length": false,
+    "no-any": true,
+    "no-consecutive-blank-lines": true,
+    "no-trailing-whitespace": true,
+    "no-duplicate-variable": true,
+    "no-var-keyword": true,
+    "no-unused-expression": true,
+    "no-use-before-declare": true,
+    "no-var-requires": true,
+    "no-require-imports": false,
+    "no-reference": false,
+    "space-before-function-paren": [true, "always"],
+    "interface-name": [true, "never-prefix"],
+    "one-line": [true,
+      "check-else",
+      "check-whitespace",
+      "check-open-brace"],
+    "quotemark": [true,
+      "single",
+      "avoid-escape"],
+    "semicolon": false,
+    "typedef-whitespace": [true, {
+      "call-signature": "nospace",
+      "index-signature": "nospace",
+      "parameter": "nospace",
+      "property-declaration": "nospace",
+      "variable-declaration": "nospace"
+    }],
+    "whitespace": [true,
+      "check-branch",
+      "check-decl",
+      "check-operator",
+      "check-separator",
+      "check-type"]
+  }
+}
