+++ packages/resolve-dependencies/typings/local.d.ts
@@ -1,0 +1,14 @@
+declare module 'replace-string' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'pnpm-install-checks' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'path-exists' {
+  const anything: any;
+  export = anything;
+}
