+++ packages/supi/package.json
@@ -30,6 +30,7 @@
     "@pnpm/package-requester": "4.1.4",
     "@pnpm/pkgid-to-filename": "2.0.0",
     "@pnpm/read-package-json": "1.0.0",
+    "@pnpm/resolve-dependencies": "0.0.0",
     "@pnpm/resolver-base": "1.2.0",
     "@pnpm/types": "^1.8.0",
     "@pnpm/utils": "0.8.0",
@@ -65,8 +66,6 @@
     "p-limit": "2.0.0",
     "package-store": "0.23.4",
     "path-absolute": "1.0.0",
-    "path-exists": "3.0.0",
-    "pnpm-install-checks": "1.1.0",
     "pnpm-shrinkwrap": "7.0.2",
     "ramda": "0.25.0",
     "remove-all-except-outer-links": "1.0.3",
@@ -77,7 +76,6 @@
     "symlink-dir": "1.1.3",
     "util.promisify": "1.0.0",
     "validate-npm-package-name": "3.0.0",
-    "version-selector-type": "2.0.0",
     "write-pkg": "3.2.0"
   },
   "devDependencies": {
@@ -102,9 +100,10 @@
     "npm-run-all": "^4.0.1",
     "npm-scripts-info": "^0.3.6",
     "package-preview": "^1.0.0",
+    "path-exists": "3.0.0",
     "path-name": "^1.0.0",
     "pnpm-registry-mock": "^2.2.0",
-    "read-pkg": "^4.0.0",
+    "read-pkg": "4.0.1",
     "rimraf": "^2.6.2",
     "sepia": "^2.0.2",
     "sinon": "^6.0.0",
