+++ packages/supi/shrinkwrap.yaml
@@ -10,6 +10,7 @@ dependencies:
   '@pnpm/package-requester': 'link:../package-requester'
   '@pnpm/pkgid-to-filename': 2.0.0
   '@pnpm/read-package-json': 1.0.0
+  '@pnpm/resolve-dependencies': 'link:../resolve-dependencies'
   '@pnpm/resolver-base': 1.2.0
   '@pnpm/types': 1.8.0
   '@pnpm/utils': 'link:../utils'
@@ -45,8 +46,6 @@ dependencies:
   p-limit: 2.0.0
   package-store: 'link:../package-store'
   path-absolute: 1.0.0
-  path-exists: 3.0.0
-  pnpm-install-checks: 1.1.0
   pnpm-shrinkwrap: 'link:../shrinkwrap'
   ramda: 0.25.0
   remove-all-except-outer-links: 1.0.3
@@ -58,7 +57,6 @@ dependencies:
   symlink-dir: 1.1.3
   util.promisify: 1.0.0
   validate-npm-package-name: 3.0.0
-  version-selector-type: 2.0.0
   write-pkg: 3.2.0
 devDependencies:
   '@pnpm/assert-project': 'link:../../privatePackages/assert-project'
@@ -82,6 +80,7 @@ devDependencies:
   npm-run-all: 4.1.3
   npm-scripts-info: 0.3.9
   package-preview: 1.0.6
+  path-exists: 3.0.0
   path-name: 1.0.0
   pnpm-registry-mock: 2.2.0
   read-pkg: 4.0.1
@@ -2827,12 +2826,6 @@ packages:
       node: '>= 0.4.0'
     resolution:
       integrity: sha1-tUGO8EOd5UJfxJlQQtztFPsqhP8=
-  /pnpm-install-checks/1.1.0:
-    dependencies:
-      semver: 5.5.1
-    dev: false
-    resolution:
-      integrity: sha1-dB2ZeXYv362T8+Rp3rSoFNNDAAg=
   /pnpm-registry-mock/2.2.0:
     dependencies:
       anonymous-npm-registry-client: 0.1.2
@@ -3927,14 +3920,6 @@ packages:
       '0': node >=0.6.0
     resolution:
       integrity: sha1-OhBcoXBTr1XW4nDB+CiGguGNpAA=
-  /version-selector-type/2.0.0:
-    dependencies:
-      semver: 5.5.1
-    dev: false
-    engines:
-      node: '>=4'
-    resolution:
-      integrity: sha512-FYMSYiI76Pd8twRTtV/DBmDkZcrMqET3ar5G+alR7/VQgo2iwnu1xlgzpvaWUCwdx5R+mq8HxgFeSnJpssV17A==
   /which/1.3.1:
     dependencies:
       isexe: 2.0.0
@@ -4064,6 +4049,7 @@ specifiers:
   '@pnpm/package-requester': 4.1.4
   '@pnpm/pkgid-to-filename': 2.0.0
   '@pnpm/read-package-json': 1.0.0
+  '@pnpm/resolve-dependencies': 0.0.0
   '@pnpm/resolver-base': 1.2.0
   '@pnpm/store-path': ^1.0.0
   '@pnpm/types': ^1.8.0
@@ -4118,11 +4104,10 @@ specifiers:
   path-absolute: 1.0.0
   path-exists: 3.0.0
   path-name: ^1.0.0
-  pnpm-install-checks: 1.1.0
   pnpm-registry-mock: ^2.2.0
   pnpm-shrinkwrap: 7.0.2
   ramda: 0.25.0
-  read-pkg: ^4.0.0
+  read-pkg: 4.0.1
   remove-all-except-outer-links: 1.0.3
   replace-string: 1.1.0
   resolve-link-target: 1.0.1
@@ -4139,7 +4124,6 @@ specifiers:
   typescript: ^3.0.0
   util.promisify: 1.0.0
   validate-npm-package-name: 3.0.0
-  version-selector-type: 2.0.0
   write-json-file: ^3.0.0
   write-pkg: 3.2.0
   write-yaml-file: ^1.0.1
