+++ packages/supi/src/unlink/index.ts
@@ -1,12 +1,11 @@
 import logger, { streamParser } from '@pnpm/logger'
 import { fromDir as readPkgFromDir } from '@pnpm/read-package-json'
-import { realNodeModulesDir } from '@pnpm/utils'
+import { getAllDependenciesFromPackage, realNodeModulesDir } from '@pnpm/utils'
 import isInnerLink = require('is-inner-link')
 import isSubdir = require('is-subdir')
 import fs = require('mz/fs')
 import path = require('path')
 import rimraf = require('rimraf-then')
-import depsFromPackage from '../depsFromPackage'
 import getContext from '../getContext'
 import { install } from '../install'
 import extendOptions, {
@@ -39,7 +38,7 @@ export async function _unlinkPkgs (
 ) {
   const modules = await realNodeModulesDir(opts.prefix)
   const pkg = await readPkgFromDir(opts.prefix)
-  const allDeps = depsFromPackage(pkg)
+  const allDeps = getAllDependenciesFromPackage(pkg)
   const packagesToInstall: string[] = []
 
   for (const pkgName of pkgNames) {
