+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
+import getAllDependenciesFromPackage from './getAllDependenciesFromPackage'
 import getSaveType from './getSaveType'
 import realNodeModulesDir from './realNodeModulesDir'
 import safeReadPackage from './safeReadPkg'
@@ -5,9 +6,13 @@ import { fromDir as safeReadPackageFromDir } from './safeReadPkg'
 import readPackage from './safeReadPkg'
 
 export {
+  getAllDependenciesFromPackage,
   getSaveType,
   readPackage,
   realNodeModulesDir,
   safeReadPackage,
   safeReadPackageFromDir,
 }
+
+export * from './nodeIdUtils'
+export * from './getWantedDependencies'
